+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexingScrubDangling.java
@@ -54,7 +54,6 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
@@ -194,7 +193,7 @@ public class IndexingScrubDangling extends IndexingBase {
                                         return false;
                                     }
                                 }
-                                return !Objects.equals(cont, rangeEnd);
+                                return !allRangesExhausted(cont, rangeEnd);
                             }));
         });
     }