+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexingScrubMissing.java
@@ -190,7 +190,7 @@ public class IndexingScrubMissing extends IndexingBase {
                                         return false;
                                     }
                                 }
-                                return !Objects.equals(cont, rangeEnd);
+                                return !allRangesExhausted(cont, rangeEnd);
                             }));
         });
     }