+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/OnlineIndexerConflictsTest.java
@@ -197,6 +197,7 @@ public class OnlineIndexerConflictsTest extends OnlineIndexerTest {
     }
 
     @Test
+    @SuppressWarnings("deprecation")
     public void testAddRecordOutsideRangeWhileIndexedIdempotent() {
 
         List<TestRecords1Proto.MySimpleRecord> records =